+++ package/src/lib.rs
@@ -42,37 +42,18 @@
 //!     .run_reverse();
 //! ```
 //!     
-//! Async version also available with `async_std` and `tokio` features:
+//! Async version also available with `async_std`, `smol` and `tokio` features:
 //!
-//! ```rust
-//! // This is a `async_std` example
-//!
+//! ```ignore
 //! use get_dir::{
 //!     GetDir,
 //!     Target,
 //!     FileTarget,
+//!     // async_std
 //!     async_std::GetDirAsyncExt,
-//! };
-//!
-//! # async fn example() {
-//! GetDir::new()
-//!     .targets(vec![
-//!         Target::File(FileTarget {
-//!             name: "LICENSE",  
-//!         }),
-//!     ])
-//!     .run_reverse_async()
-//!     .await;
-//! # }
-//! ```
-//!
-//! ```rust
-//! // This is a `tokio` example
-//!
-//! use get_dir::{
-//!     GetDir,
-//!     Target,
-//!     FileTarget,
+//!     // smol
+//!     smol::GetDirAsyncExt,
+//!     // tokio
 //!     tokio::GetDirAsyncExt,
 //! };
 //!
@@ -98,6 +79,17 @@
 /// ```
 #[cfg(feature = "async_std")]
 pub mod async_std;
+
+/// Run asynchronously with `smol` feature.
+///
+/// To use it, add the following code to the `Cargo.toml` file:
+///
+/// ```toml
+/// [dependencies]
+/// get_dir = { version = "*", features = ["smol"] }
+/// ```
+#[cfg(feature = "smol")]
+pub mod smol;
 
 /// Run asynchronously with `tokio` feature.
 ///